pip-requirements.txt
networkx
matplotlib
biopython
numpy